WEP0, a proposal for the WEP process
This is the first Write the Docs Enhancement Proposal (WEP), which introduces the WEP process itself and proposes to implement it in the Write the Docs community. The WEP process enables anyone from the community to propose process and policy changes to the community. The community then reviews the proposed change and if the change is accepted, the policy or process change is implemented in the community.

This "meta-WEP" follows the WEP process to show how the process works and to help the community learn how to use it.

Please read below and comment.